Post-meeting commit
Now with method of adjustment trials. Although in the Quest (it works in the Quest!) I realized that pressing the button sliiiightly too long would cause the books to shuffle, which is not comfortable to look at (and records confirmed trials by accident!) So I will fix that. I will also fix the scaling of the bookshelf so it is not huge (just very far away that's why you think it looks normal.)